refactor api_fetch_extractor to own file
